feat: Setup deployment via Docker
Adds a Dockerfile. Nginx is used as the base image, and envsubst lets us
inject the Docker environment variables into Nginx's config, which can
then inject them into the page. This means that it is now possible to
either run the build process manually (which will use the settings from
config.js) or use the Docker image (which will use environment
variables, insofar as they are defined by the user).