chore: Avoid duplicating page metadata in HTML and manifest (#139)
This patch adds a custom Vite plugin to the config. The plugin takes the
PWA manifest definition (and a bit extra) as input, and generates the
appropriate link and meta tags from that. These things can then be
removed from the HTML. Hence, there is only a single place where they
are defined instead of redundancy that could get out of sync.