chore: Proxy API requests during development to remove CORS headers (#242)
The API server may be configured to only allow a specific CORS origin.
This interferes with development on the `locahost` origin, which doesn't
match the CORS origin. To solve this, the dev server will now proxy API
requests to remove any CORS limitations. Note that production usage will
not be affected by this change.